[F] Fixed int time

+++ Runtime/TimerRaiser/Domain/ITimer.cs
@@ -5,6 +5,6 @@ namespace TimerRaiser
     public interface ITimer
     {
         Action OnTimerDone { get; set; }
-        void StartTimer(int time);
+        void StartTimer(float time);
     }
 }
